requirements.txt
asttokens==2.4.1
boto3==1.34.11
botocore==1.34.11
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
executing==2.0.1
greenlet==3.0.3
idna==3.6
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.0
jupyter_core==5.6.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
SQLAlchemy==2.0.24
stack-data==0.6.3
tabula-py==2.9.0
tornado==6.4
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.12
